+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - logger: use `_location` and `_module` instead of `@location` and `@module`.
 - logger: remove the cargo prefix from locations.
 - telemeter: hide rendered metrics in dumps.
+- dumping: responses are dumped with the `RequestName::Response` name.
 
 ## [0.1.27] - 2021-09-27
 ### Fixed